Testing: Extracted copy tests to their own class
